% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_background_scores}
\alias{get_background_scores}
\title{get_background_scores}
\usage{
get_background_scores()
}
\value{
Data frame with match scores per phosphosite (rows) and PWMs (columns)
}
\description{
A data frame with precomputed PWM scores for a large set of curated
phosphosites
}
\examples{
bg <- get_background_scores()
}
