% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_kinase_annotation}
\alias{get_kinase_annotation}
\title{get_kinase_annotation}
\usage{
get_kinase_annotation()
}
\value{
A data frame with columns MatrixName, GeneName, UniprotID, EntrezID, Description and KinaseFamily
}
\description{
Annotation data for all 303 human kinase PWMs
}
\examples{
anno <- get_kinase_annotation()
}
