% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_phosphosites.R
\name{get_kinase_pwms}
\alias{get_kinase_pwms}
\title{get_kinase_pwms}
\usage{
get_kinase_pwms(include_ST_favorability = TRUE)
}
\arguments{
\item{include_ST_favorability}{Include serine vs. threonine favorability for
the central phospho-acceptor?}
}
\value{
A named list of numeric matrices (PWMs).
}
\description{
Get a list of position weight matrices (PWMs) for the 303 human
serine/threonine kinases originally published in Johnson et al. 2023.
}
\examples{
pwms <- get_kinase_pwms()
}
