% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getBackgroundScores}
\alias{getBackgroundScores}
\title{Get precomputed PWM scores for a large set of curated human phosphosites}
\usage{
getBackgroundScores()
}
\value{
A data frame with log2-odds scores per phosphosite (rows) and PWMs
(columns)
}
\description{
Get precomputed PWM scores for a large set of curated human phosphosites
}
\examples{
bg <- getBackgroundScores()
}
