% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getKinaseAnnotation}
\alias{getKinaseAnnotation}
\title{Get annnotation data for all 303 human serine/threonine kinase PWMs}
\usage{
getKinaseAnnotation()
}
\value{
A data frame with columns MatrixName, GeneName, UniprotID, EntrezID,
Description and KinaseFamily
}
\description{
Get annnotation data for all 303 human serine/threonine kinase PWMs
}
\examples{
anno <- getKinaseAnnotation()
}
