% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_phosphosites.R
\name{process_phosphosites}
\alias{process_phosphosites}
\title{Process phosphosites to a common format used for PWM matching}
\usage{
process_phosphosites(sites, allow_phospho_priming = TRUE)
}
\arguments{
\item{sites}{Character vector with phosphosites}

\item{allow_phospho_priming}{Allow phospho-acceptors at non-central
positions? These should be indicated by the lower case letters "s", "t" or
"y".}
}
\value{
A tibble with columns: \code{sites}, \code{processed}, \code{residue}
}
\description{
The central phospho-acceptor of each phosphosite is recognized in two ways.
Either the acceptor residue is followed by an asterisk (\emph{) character, e.g.,
"SAGLLS}DEDC". Alternatively, it is defined as the central residue i.e. at
position \code{floor(width(site))+1}.
}
\details{
Non-central phospho-acceptors (phospho-priming) should be indicated by the
lower case letters "s", "t" or "y". If phospho-priming is disabled these
residues are converted to upper case letters.

The input sites are truncated and/or padded such that the processed sites are
of width 10 and have the central phospho-acceptor surrounded by 5 upstream
and 4 downstream residues.

A warning is raised if the central phospho-acceptor is not serine or
threonine, as these sites are not covered by the Johnson PWMs.
}
\examples{
proc_sites <- process_phosphosites(c("SAGLLS*DEDC", "GDS*ND", "EKGDSN__", "___LySDEDC", "EKGtS*N"))
}
