% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_phosphopeptides.R
\name{processPhosphosites}
\alias{processPhosphosites}
\title{Process phospho-peptides to a common format used for PWM matching}
\usage{
processPhosphosites(
  sites,
  onlyCentralAcceptor = TRUE,
  allowPhosphoPriming = TRUE
)
}
\arguments{
\item{sites}{Character vector with phospho-peptides}

\item{allowPhosphoPriming}{Allow phospho-acceptors at non-central positions?
These should be indicated by the lower case letters "s", "t" or "y".}
}
\value{
A tibble with columns: \code{sites}, \code{processed}, \code{residue}
}
\description{
Phosphorylated residues (=phospho-acceptors) are recognized either by lower
case letters (s, t or y) or the phosphorylated residue is followed by an
asterisk (S*,T*, Y*).
}
\details{
(1) Only the central phospho-acceptor of each phospho-peptide is used. It is
defined as the one left-closest to position \code{floor(width(site)/2)+1}. (2) All

All other lower case "s"/"t"/"y" are treated as non-central phospho-acceptors
that enable phospho-priming. If phospho-priming is disabled (parameter
\code{allowPhosphoPriming}) these residues are converted to upper case letters.

The input sites are truncated and/or padded such that the processed sites are
of width 10 and have the central phospho-acceptor surrounded by 5 upstream
and 4 downstream residues.

A warning is raised if the central phospho-acceptor is not serine or
threonine, as these sites are not covered by the Johnson PWMs.
}
\examples{
procSites <- processPhosphosites(c("SAGLLS*DEDC", "GDS*ND", "EKGDSN__", "___LySDEDC", "EKGtS*N"))
}
