% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_phosphosites.R
\name{score_phosphosites}
\alias{score_phosphosites}
\title{Match kinase PWMs to processed phosphosites}
\usage{
score_phosphosites(
  pwms,
  sites,
  score_type = c("percentile", "log2_odds"),
  BPPARAM = BiocParallel::MulticoreParam(1)
)
}
\arguments{
\item{pwms}{List with kinase PWMs as returned by \code{\link[=get_kinase_pwms]{get_kinase_pwms()}}.}

\item{sites}{A character vector with phosphosites. Check
\code{\link[=process_phosphosites]{process_phosphosites()}} for the correct phosphosite format.}

\item{score_type}{Percentile rank or log2-odds score.}

\item{BPPARAM}{A BiocParallelParam object specifying how parallelization
should be performed.}
}
\value{
A numeric matrix of size \code{length(sites)} times \code{length(kinases)}.
}
\description{
\code{score_phosphosites} takes a list of kinase PWMs and a vector of processed
phosphosites as input and returns a matrix of match scores per PWM and site.
}
\details{
The score is either the PWM match score (\code{log2_odds}) or the percentile rank
(\code{percentile}) in the background score distribution.
}
\examples{
score <- score_phosphosites(get_kinase_pwms(), c("TGRRHTLAEV", "LISAVSPEIR"))
}
\seealso{
\code{\link[=get_kinase_pwms]{get_kinase_pwms()}} for getting a list of kinase PWMs,
\code{\link[=process_phosphosites]{process_phosphosites()}} for the correct phosphosite format, and
\code{\link[=get_score_maps]{get_score_maps()}} for mapping PWM scores to percentile ranks
}
