% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getKinaseAnnotation}
\alias{getKinaseAnnotation}
\title{Get annnotation data for all 303 human serine/threonine kinase PWMs}
\usage{
getKinaseAnnotation()
}
\value{
A data frame with columns MatrixName, GeneName, UniprotID, EntrezID,
Description and KinaseFamily
}
\description{
The annotation data records for each of the 303 human serine/threonine kinase
PWMs originally published in Johnson et al. the PWM matrix name, gene symbol
and description, Uniprot ID and Entrez ID as well as the kinase family.
}
\examples{
anno <- getKinaseAnnotation()
}
