% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_phosphosites.R
\name{getKinasePWM}
\alias{getKinasePWM}
\title{Get a list of position weight matrices (PWMs) for the 303 human
serine/threonine kinases originally published in Johnson et al. 2023.}
\usage{
getKinasePWM(includeSTfavorability = TRUE)
}
\arguments{
\item{includeSTfavorability}{Include serine vs. threonine favorability for
the central phospho-acceptor?}
}
\value{
A named list of numeric matrices (PWMs).
}
\description{
Get a list of position weight matrices (PWMs) for the 303 human
serine/threonine kinases originally published in Johnson et al. 2023.
}
\examples{
pwms <- getKinasePWM()
}
