% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JohnsonKinaseData-package.R
\docType{package}
\name{JohnsonKinaseData-package}
\alias{JohnsonKinaseData}
\alias{JohnsonKinaseData-package}
\title{JohnsonKinaseData: Kinase PWMs based on data published by Johnson et al. 2023 and Yaron-Barir et al. 2024}
\description{
The packages provides position specific weight matrices (PWMs) for 303 human serine/threonine and 93 tyrosine kinases originally published in Johnson et al. 2023 (doi:10.1038/s41586-022-05575-3) and Yaron-Barir et al. 2024 (doi:10.1038/s41586-024-07407-y). The package includes basic functionality to score user provided phosphosites. It also includes pre-computed PWM scores ("background scores") for a large collection of curated human phosphosites which can be used to rank PWM scores relative to the background scores ("percentile rank").
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fgeier/JohnsonKinaseData/}
  \item Report bugs at \url{https://support.bioconductor.org/t/JohnsonKinaseData}
}

}
\author{
\strong{Maintainer}: Florian Geier \email{florian.geier@unibas.ch} (\href{https://orcid.org/0000-0002-9076-9264}{ORCID})

}
\keyword{internal}
