% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JohnsonKinaseData-package.R
\docType{package}
\name{JohnsonKinaseData-package}
\alias{JohnsonKinaseData-package}
\alias{JohnsonKinaseData}
\title{JohnsonKinaseData}
\description{
The \pkg{JohnsonKinaseData} package provides position specific
weight matrices (PWMs) for 303 human serine/threonine kinases originally
published in Johnson et al. 2023.

It also includes pre-computed PWM scores ("background scores") for a large
collection of curated human phosphosites.

The package additionally offers functionality to match kinase PWMs against
user-provided phospho-peptides and to rank PWM scores relative to the
background scores ("percentile rank").
}
\references{
Johnson, J.L., Yaron, T.M., Huntsman, E.M. et al. An atlas of
substrate specificities for the human serine/threonine kinome. Nature 613,
759–766 (2023). https://doi.org/10.1038/s41586-022-05575-3
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fgeier/JohnsonKinaseData/}
  \item Report bugs at \url{https://support.bioconductor.org/t/JohnsonKinaseData}
}

}
\author{
Florian Geier (florian.geier@unibas.ch)
}
